sed fixes in GHC classes
xenclient-nws was failing to build due to the OpenXT ghc based libraries
trying to reach outside of the ndvm sysroot during linking.  This was
traced back to incomplete clean up of the sysroot path being in ghc libs
library and include paths.  These are in the lib's conf file.  Fixed the
classes to properly strip out the sysroot on do_install and properly add
the correct sysroot when the conf is pulled into another recipe for
compilation.